+++ src/Pgsql/Configuration/PgSQLActionConfigRowDefinition.php
@@ -16,6 +16,7 @@ class PgSQLActionConfigRowDefinition implements ConfigurationInterface
         $rootNode = $treeBuilder->getRootNode();
 
         $rootNode
+            ->ignoreExtraKeys(false)
             ->children()
                 ->scalarNode('data_dir')
                     ->isRequired()
