+++ tests/data/functionalRow/config.json
@@ -1,0 +1,47 @@
+{
+  "parameters": {
+    "db": {
+      "driver": "pgsql",
+      "host": "pgsql",
+      "port": "5432",
+      "database": "postgres",
+      "schema": "public",
+      "user": "postgres",
+      "password": "some password",
+      "#password": "some password"
+    },
+    "tableId": "simple",
+    "dbName": "simple",
+    "export": true,
+    "incremental": false,
+    "primaryKey": [
+      "id"
+    ],
+    "items": [
+      {
+        "name": "id",
+        "dbName": "id",
+        "type": "int",
+        "size": null,
+        "nullable": null,
+        "default": null
+      },
+      {
+        "name": "name",
+        "dbName": "name",
+        "type": "varchar",
+        "size": 255,
+        "nullable": null,
+        "default": null
+      },
+      {
+        "name": "glasses",
+        "dbName": "glasses",
+        "type": "varchar",
+        "size": 255,
+        "nullable": null,
+        "default": null
+      }
+    ]
+  }
+}
