+++ tests/data/functionalRow/in/tables/simple.csv
@@ -1,0 +1,10 @@
+"id","name","glasses"
+"0","Miro cillik","yes"
+"1","Ondrej Hlavacek","no"
+"2","Martin Halamicek","yes"
+"3","Tomas Kacur","yes"
+"4","Erik Zigo","no"
+"5","Marc Raiser","sometimes"
+"6","Petr Simecek","yes"
+"7","Jakub Matejka","yes"
+"8","Ondrej Popelka","no"
